requirements.txt
flask
transformers
einops
torchvision
Pillow